Update RequestController.java
Organized the flow to check for Method Type first and then based on that check for parameters and headers and put them in the URL if it's GET , put them in the body if it's PUT OR POST. Some portions are still in the todo stage.  This is actually pre-header cleanup and organization work.